requirements.txt
langchain-cohere==0.4.2
langchain-chroma==0.2.2
python-dotenv==1.0.1
langchain-community==0.3.18